requirements.txt
python-telegram-bot>=20.0,<23.0
python-dotenv>=1.0,<1.1
pandas>=2.0,<2.1
urllib3>=1.26.15,<2.0
